add: redLines Textfield, remove: texthelper
